Add utils module and normalize functions
