+++ data.py
@@ -1,14 +1,10 @@
 #
 # gets all builds for all heroes (prints or writes to file (json?)
-from multiple import get_builds, print_build
+from utils import get_soup, normalize_hero_name
+from scrapping import get_hero_builds, print_build
 import json
 import requests
 from bs4 import BeautifulSoup
-
-
-def get_soup(link):
-    page = requests.get(link)
-    return BeautifulSoup(page.content, 'html.parser')
 
 
 '''
@@ -22,9 +18,11 @@ def update_heroes_list():
     for role in roles:
         link = 'https://www.icy-veins.com/heroes/' + role + '-hero-guides'
         soup = get_soup(link)
+
         hero_block_tags = soup.find_all('div', class_='nav_content_block_entry_heroes_hero')
         for hero_block in hero_block_tags:
-            heroes_names.add(hero_block.find('span', class_='').contents[0])
+            name = normalize_hero_name(hero_block.find('span', class_='').contents[0])
+            heroes_names.add(name)
 
     with open('data/heroes.json', 'w') as fp:
         json.dump(list(heroes_names), fp)
@@ -33,9 +31,10 @@ def update_heroes_list():
 '''
 Updates list of builds (in [hero name].json) for a given hero
 '''
-def update_builds(hero):
+def update_hero_builds(hero):
+    hero = normalize_hero_name(hero)
     print('Updating builds for ' + hero + '...', end='')
-    builds, titles = get_builds(hero)
+    builds, titles = get_hero_builds(hero)
     filename = 'data/builds/' + hero + '.json'
 
     with open(filename, 'w') as fp:
@@ -55,16 +54,19 @@ def update_all_builds():
         heroes = json.load(fp)
 
     for hero in heroes:
-        update_builds(hero)
+        update_hero_builds(hero)
 
 
 '''
 Loads builds for one hero
 '''
 def load_builds(hero):
+    hero = normalize_hero_name(hero)
     filename = 'data/builds/' + hero + '.json'
+
     with open(filename, 'r') as fp:
         data = json.load(fp)
+
     print(data)
 
 
@@ -82,3 +84,4 @@ def print_all_builds():
         for build, title in zip(builds, titles):
             print_build(levels, build, title)
 
+update_all_builds()
